some changes

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, String, Float, Date
+from sqlalchemy import Column, Integer, ForeignKey, String, Date
 from sqlalchemy.orm import relationship, declarative_base
 
 Base = declarative_base()
@@ -29,7 +29,7 @@ class RezeptZutat(Base):
     __tablename__ = "rezept_zutaten"
     rezept_id = Column(Integer, ForeignKey("rezepte.id"), primary_key=True)
     zutat_id = Column(Integer, ForeignKey("zutaten.id"), primary_key=True)
-    menge = Column(Float, nullable=False)
+    menge = Column(Integer, nullable=False)
 
     rezept = relationship("Rezept", back_populates="rezept_zutaten")
     zutat = relationship("Zutat", back_populates="rezept_zutaten")
